+++ src/idunno.Bluesky/Notifications/Notification.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Barry Dorrans. All rights reserved.
 // Licensed under the MIT License.
 
+using System.Collections.ObjectModel;
 using System.Text.Json.Serialization;
 
 using idunno.AtProto;
@@ -24,6 +25,7 @@ public sealed record Notification : AtProtoRepositoryObject
         notificationResponse.Cid,
         notificationResponse.Author,
         notificationResponse.Reason,
+        notificationResponse.ReasonSubject,
         notificationResponse.Record,
         notificationResponse.StarterPack,
         notificationResponse.IsRead,
@@ -35,8 +37,9 @@ public sealed record Notification : AtProtoRepositoryObject
     internal Notification(
         AtUri uri,
         Cid cid,
-        ProfileViewBasic author,
+        ProfileView author,
         string reason,
+        AtUri? reasonSubject,
         BlueskyRecord record,
         StarterPackViewBasic? starterPack,
         bool isRead,
@@ -45,19 +48,13 @@ public sealed record Notification : AtProtoRepositoryObject
     {
         Author = author;
         RawReason = reason;
+        ReasonSubject = reasonSubject;
         Record = record;
         IsRead = isRead;
         IndexedAt = indexedAt;
         StarterPack = starterPack;
 
-        if (labels is not null)
-        {
-            Labels = labels;
-        }
-        else
-        {
-            Labels = new List<Label>().AsReadOnly<Label>();
-        }
+        Labels = labels!;
 
         // As JsonStringEnumConvertor doesn't have the ability to fall back to a default, we have to do this.
         Reason = reason.ToUpperInvariant() switch
@@ -74,16 +71,17 @@ public sealed record Notification : AtProtoRepositoryObject
             "LIKE-VIA-REPOST" => NotificationReason.LikeViaRepost,
             "REPOST-VIA-REPOST" => NotificationReason.RepostViaRepost,
             "SUBSCRIBED-POST" => NotificationReason.SubscribedPost,
+            "CONTACT-MATCH" => NotificationReason.ContactMatch,
             _ => NotificationReason.Unknown,
         };
     }
 
     /// <summary>
-    /// Gets the <see cref="ProfileViewBasic"/> of the actor that triggered the notification.
+    /// Gets the <see cref="ProfileView"/> of the actor that triggered the notification.
     /// </summary>
     [JsonInclude]
     [JsonRequired]
-    public ProfileViewBasic Author { get; init; }
+    public ProfileView Author { get; init; }
 
     /// <summary>
     /// Gets the raw reason for the notification.
@@ -98,11 +96,10 @@ public sealed record Notification : AtProtoRepositoryObject
     /// </summary>
     /// <remarks><para>If the reason is <see cref="NotificationReason.Unknown"/> the unparsed reason can be accessed via <see cref="RawReason"/>.</para></remarks>
     [JsonIgnore]
-    [JsonPropertyName("reason_enum")]
     public NotificationReason Reason { get; init; }
 
     /// <summary>
-    /// Gets the <see cref="AtUri"/> of the subject that triggered the notification. if any.
+    /// Gets the <see cref="AtUri"/> of the subject that triggered the notification, if any.
     /// </summary>
     [JsonInclude]
     public AtUri? ReasonSubject { get; init; }
@@ -136,5 +133,9 @@ public sealed record Notification : AtProtoRepositoryObject
     /// Gets a collection of <see cref="Label"/>s for the notification.
     /// </summary>
     [JsonInclude]
-    public IReadOnlyCollection<Label> Labels { get; init; }
+    public IReadOnlyCollection<Label> Labels
+    {
+        get;
+        init => field = value is null ? ReadOnlyCollection<Label>.Empty : new List<Label>(value).AsReadOnly();
+    }
 }